apps/api/pyproject.toml
[project]
name = "ai-trainer-api"
version = "0.1.0"
description = "AI Training Platform for Sales Managers - Backend API"
requires-python = ">=3.11"
dependencies = [
    # Core web framework — pin below 1.0 until migration guide available
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "alembic>=1.14.0,<2.0.0",
    # Validation
    "pydantic>=2.10.0,<3.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    # Auth (PyJWT — actively maintained; migrated from python-jose)
    "PyJWT>=2.8.0,<3.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt>=4.0.0,<5.0.0",
    # Redis / HTTP
    "redis>=5.2.0,<6.0.0",
    "httpx>=0.28.0,<1.0.0",
    # LLM providers
    "anthropic>=0.40.0,<1.0.0",
    "openai>=1.58.0,<2.0.0",
    # Utilities
    "python-multipart>=0.0.12,<1.0.0",
    "pgvector>=0.3.6,<1.0.0",
    "aiofiles>=24.1.0,<25.0.0",
    "slowapi>=0.1.9,<1.0.0",
    "aiosmtplib>=3.0.0,<4.0.0",
    "socksio>=1.0.0,<2.0.0",
    "Pillow>=10.0.0,<12.0.0",
    "gunicorn>=22.0.0,<24.0.0",
    # PDF generation (lightweight, pure Python)
    "fpdf2>=2.8.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
